[determinism] Add softmax/xent GPU-determinism.
